refactor(workbench-testing-app): move application config to separate file
